+++ HISTORY.md
@@ -26,7 +26,8 @@
     * German
 
 ### 2025-12-06
-* gui: exit the tutorial window when creating or opening a new project 
+* gui: exit the tutorial window when creating or opening a new project
+* gui: make sure to switch to a non-tutorial scene when launching a tutorial
 
 ### 2025-12-05
 * gui, extensions: prevent programmatic downscaling of the tutorial window below its minimum size